Actualización del atributo CapacidadWatts a su tipo de dato correspondiente

+++ backend/ArmatuXPC.Backend/DTOS/Componentes/ComponenteDto.cs
@@ -8,7 +8,7 @@ public class ComponenteDto
     public decimal Precio { get; set; }
     public TipoComponente Tipo { get; set; }
     public decimal? ConsumoWatts { get; set; }
-    public int? CapacidadWatts { get; set; }
+    public decimal? CapacidadWatts { get; set; }
     public string? ImagenUrl { get; set; }
     public bool EstaActivo {get; set; }
 }
